+++ database/factories/SeriesFactory.php
@@ -1,8 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DrewRoberts\Blog\Database\Factories;
 
-use App\Models\User;
 use DrewRoberts\Blog\Models\Series;
 use DrewRoberts\Blog\Models\Topic;
 use Illuminate\Support\Str;
@@ -23,8 +24,8 @@ class SeriesFactory extends Factory
             'description'       => $this->faker->sentences(1, true),
             'pageviews'         => $this->faker->numberBetween(0, 5000),
             'topic_id'          => randomOrCreate(Topic::class),
-            'creator_id'        => randomOrCreate(User::class),
-            'updater_id'        => randomOrCreate(User::class),
+            'creator_id'        => randomOrCreate(app('user')),
+            'updater_id'        => randomOrCreate(app('user'))
         ];
     }
 }