+++ database/factories/TopicFactory.php
@@ -1,8 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DrewRoberts\Blog\Database\Factories;
 
-use App\Models\User;
 use DrewRoberts\Blog\Models\Topic;
 use Illuminate\Support\Str;
 use Illuminate\Database\Eloquent\Factories\Factory;
@@ -21,8 +22,8 @@ class TopicFactory extends Factory
             'description'       => $this->faker->sentences(1, true),
             'note'              => $this->faker->sentences(1, true),
             'pageviews'         => $this->faker->numberBetween(1, 400),
-            'creator_id'        => randomOrCreate(User::class),
-            'updater_id'        => randomOrCreate(User::class),
+            'creator_id'        => randomOrCreate(app('user')),
+            'updater_id'        => randomOrCreate(app('user'))
         ];
     }
 }