+++ database/migrations/2017_05_20_110000_create_series_table.php
@@ -18,13 +18,13 @@ class CreateSeriesTable extends Migration
             $table->string('ogdescription')->nullable(); // Open Graph Description used for social shares. Will default to description if NULL.
             $table->unsignedInteger('pageviews')->index(); // Total current pageviews for series page. Will be synced from Google Analytics API.
 
-            $table->foreignId('topic_id')->references('id')->on('topics'); // Group series into larger topics
-            $table->foreignId('image_id')->nullable()->references('id')->on('images'); // Cover image for series
-            $table->foreignId('ogimage_id')->nullable()->references('id')->on('images'); // External open graph image id. Featured image for social sharing. Will default to image_id unless this is used. Allows override for play button or words on image.
-            $table->foreignId('video_id')->nullable()->references('id')->on('videos'); // If series has a featured video.
+            $table->foreignIdFor(app('topic')); // Group series into larger topics
+            $table->foreignIdFor(app('image'))->nullable(); // Cover image for series
+            $table->foreignIdFor(app('image'), 'ogimage_id')->nullable(); // External open graph image id. Featured image for social sharing. Will default to image_id unless this is used. Allows override for play button or words on image.
+            $table->foreignIdFor(app('video'))->nullable(); // If series has a featured video.
 
-            $table->foreignId('creator_id')->references('id')->on('users');
-            $table->foreignId('updater_id')->references('id')->on('users');
+            $table->foreignIdFor(app('user'), 'creator_id');
+            $table->foreignIdFor(app('user'), 'updater_id');
             $table->timestamps();
         });
     }