+++ src/BlogServiceProvider.php
@@ -1,35 +1,25 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DrewRoberts\Blog;
 
-use DrewRoberts\Blog\Commands\BlogCommand;
-use Illuminate\Support\ServiceProvider;
+use Spatie\LaravelPackageTools\Package;
+use Spatie\LaravelPackageTools\PackageServiceProvider;
 
-class BlogServiceProvider extends ServiceProvider
+class BlogServiceProvider extends PackageServiceProvider
 {
     public function boot()
     {
-        $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
+        $this->loadMigrationsFrom(__DIR__ . '/../database/migrations');
 
-        if ($this->app->runningInConsole()) {
-            $this->publishes([
-                __DIR__.'/../config/blog.php' => config_path('blog.php'),
-            ], 'config');
-
-            $this->publishes([
-                __DIR__.'/../resources/views' => base_path('resources/views/vendor/blog'),
-            ], 'views');
-
-            $this->commands([
-                BlogCommand::class,
-            ]);
-        }
-
-        $this->loadViewsFrom(__DIR__.'/../resources/views', 'blog');
+        parent::boot();
     }
 
-    public function register()
+    public function configurePackage(Package $package): void
     {
-        $this->mergeConfigFrom(__DIR__.'/../config/blog.php', 'blog');
+        $package
+            ->name('blog')
+            ->hasConfigFile();
     }
 }