+++ src/Models/Page.php
@@ -1,21 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DrewRoberts\Blog\Models;
 
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use Tipoff\Support\Traits\HasPackageFactory;
+use Tipoff\Support\Models\BaseModel;
 use Tipoff\Support\Traits\HasCreator;
+use Tipoff\Support\Traits\HasPackageFactory;
 use Tipoff\Support\Traits\HasUpdater;
 
-class Page extends Model
+class Page extends BaseModel
 {
-    use SoftDeletes;
-    use HasPackageFactory;
-    use HasCreator;
-    use HasUpdater;
+
+    use SoftDeletes, HasCreator, HasUpdater, HasPackageFactory;
 
     protected $guarded = ['id'];
 
@@ -26,8 +26,7 @@ class Page extends Model
     protected static function boot()
     {
         parent::boot();
-
-
+      
         static::saving(function ($page) {
             if (empty($page->author_id)) { // Can specify a different author for a page than Auth user
                 $page->author_id = auth()->user()->id;